% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_coefficient.R
\name{adjust_coef}
\alias{adjust_coef}
\alias{adjust_coef_with_continuous}
\title{Adjust an observed regression coefficient for a normally distributed
confounder}
\usage{
adjust_coef(
  effect_observed,
  exposure_confounder_effect,
  confounder_outcome_effect,
  verbose = getOption("tipr.verbose", TRUE)
)

adjust_coef_with_continuous(
  effect_observed,
  exposure_confounder_effect,
  confounder_outcome_effect,
  verbose = getOption("tipr.verbose", TRUE)
)
}
\arguments{
\item{effect_observed}{Numeric. Observed exposure - outcome effect from a regression
model. This can be the beta coefficient, the lower confidence bound of
the beta coefficient, or the upper confidence bound of the beta
coefficient.}

\item{exposure_confounder_effect}{Numeric. Estimated difference in scaled means between the
unmeasured confounder in the exposed population and unexposed population}

\item{confounder_outcome_effect}{Numeric. Estimated relationship
between the unmeasured confounder and the outcome.}

\item{verbose}{Logical. Indicates whether to print informative message.
Default: \code{TRUE}}
}
\value{
Data frame.
}
\description{
Adjust an observed regression coefficient for a normally distributed
confounder
}
\examples{
## Update an observed coefficient of 0.5 with an unmeasured confounder
## with a difference in scaled means between exposure groups of 0.2
## and coefficient of 0.3
adjust_coef(0.5, 0.2, 0.3)
}
