% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_coefficient.R
\name{adjust_rr_with_binary}
\alias{adjust_rr_with_binary}
\title{Adjust an observed relative risk with a binary confounder}
\usage{
adjust_rr_with_binary(
  effect,
  exposed_p,
  unexposed_p,
  outcome_effect,
  verbose = TRUE
)
}
\arguments{
\item{effect}{Numeric positive value. Observed exposure - outcome relative risk.
This can be the point estimate, lower confidence bound, or upper
confidence bound.}

\item{exposed_p}{Numeric between 0 and 1. Estimated prevalence of the
unmeasured confounder in the exposed population}

\item{unexposed_p}{Numeric between 0 and 1. Estimated prevalence of the
unmeasured confounder in the unexposed population}

\item{outcome_effect}{Numeric positive value. Estimated relationship
between the unmeasured confounder and the outcome}

\item{verbose}{Logical. Indicates whether to print informative message.
Default: \code{TRUE}}
}
\value{
Data frame.
}
\description{
Adjust an observed relative risk with a binary confounder
}
\examples{
adjust_rr_with_binary(1.1, 0.5, 0.3, 1.3)
}
