% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_value.R
\name{e_value}
\alias{e_value}
\title{Calculate an E-value}
\usage{
e_value(effect_observed)
}
\arguments{
\item{effect_observed}{Numeric positive value. Observed exposure - outcome effect
(assumed to be the exponentiated coefficient, so a risk ratio, odds
ratio, or hazard ratio). This can be the point estimate, lower confidence
bound, or upper confidence bound.}
}
\value{
Numeric value
}
\description{
Calculate an E-value
}
\examples{
e_value(0.9)
e_value(1.3)
}
