% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exdata_continuous.R
\docType{data}
\name{exdata_continuous}
\alias{exdata_continuous}
\title{Example Data (Continuous Outcome)}
\format{
A data frame with 2,000 rows and 4 columns:
\itemize{
\item \code{.unmeasured_confounder}: A simulated unmeasured confounder
\item \code{measured_confounder}: A simulated measured confounder
\item \code{exposure}
\item \code{outcome}
}
}
\usage{
exdata_continuous
}
\description{
A data set simulated with two Normally distributed confounders, one
"measured" and one "unmeasured", an exposure, and outcome. The "true" causal
effect of the exposure on the outcome, accounting for both the measured
and unmeasured confounders, should be 0.
}
\keyword{datasets}
