% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{expression}
\alias{expression}
\title{expression matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 19003 rows and 5 columns.
}
\usage{
expression
}
\description{
A simple expression matrix with five sample.
}
\examples{
data("expression", package="TIPR")
head(expression)


}
\keyword{datasets}
