% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{indicatorInsideGeneSetMatrix}
\alias{indicatorInsideGeneSetMatrix}
\title{identical gene position}
\usage{
indicatorInsideGeneSetMatrix(superposition_rank, gSet_pos_matrix)
}
\arguments{
\item{superposition_rank}{A matrix indicating rank of all values in the raw matrix.}

\item{gSet_pos_matrix}{A matrix consist of '0' and '1' show the position of signature genes in expression profile,
with same size of expression profile, '1' stands for signature genes expression, '0' for other genes.}
}
\value{
matrix
}
\description{
With the rank of all values in the matrix(superposition.rank),
get the matrix consist with 0 and 1, indicated the position of geneset, required for calculating ES score of the geneset.
}
