% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed_bias_order.R
\name{observed_bias_order}
\alias{observed_bias_order}
\title{Order observed bias data frame for plotting}
\usage{
observed_bias_order(d, by)
}
\arguments{
\item{d}{Observed bias data frame. Must have columns \code{dropped} and \code{group}}

\item{by}{Character. Variable in \code{d} to order by.}
}
\value{
Data frame in the correct order
}
\description{
Order observed bias data frame for plotting
}
