% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed_covariate_e_value.R
\name{observed_covariate_e_value}
\alias{observed_covariate_e_value}
\title{Calculate the Observed Covariate E-value}
\usage{
observed_covariate_e_value(lb, ub, lb_adj, ub_adj, transform = NULL)
}
\arguments{
\item{lb}{Numeric. The lower bound of the full model}

\item{ub}{Numeric. The upper bound of the full model}

\item{lb_adj}{Numeric. The lower bound of the adjusted model}

\item{ub_adj}{Numeric. The upper bound of the adjusted model}

\item{transform}{Character. If your effect is an odds ratio or hazard ratio, this will
perform the transformation suggested by VanderWeele and Ding. Allowed values are:
\itemize{
\item "OR"
\item "HR"
}}
}
\value{
The Observed Covariate E-value
}
\description{
Calculate the Observed Covariate E-value
}
