% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sortMatrixByCol}
\alias{sortMatrixByCol}
\title{sort matrix by column}
\usage{
sortMatrixByCol(raw_matrix, decreasing = TRUE)
}
\arguments{
\item{raw_matrix}{A matrix consist of numeric value.}

\item{decreasing}{Logical, if true sort is decreasing, default by TRUE.}
}
\description{
Sort every column of a matrix.
}
