% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ssGSEA_ES}
\alias{ssGSEA_ES}
\title{ES score}
\usage{
ssGSEA_ES(exp_rank, superposition_rank, gSet_pos_matrix, alpha = 0.25)
}
\arguments{
\item{exp_rank}{A matrix indicating rank of every element in column.}

\item{superposition_rank}{A matrix indicating rank of all values in the matrix.}

\item{gSet_pos_matrix}{A matrix consist of '0' and '1' show the position of signature genes in expression profile,
with same size of expression profile, '1' stands for signature genes expression, '0' for other genes.}

\item{alpha}{Quantity alpha usually set to 0.25.}
}
\value{
vector of numeric values indicating ES score of one geneset for every sample
}
\description{
Calculating ES score of one geneset.
}
