% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_tip.R
\name{tidy_tip}
\alias{tidy_tip}
\title{Tidy tipr}
\usage{
tidy_tip(mod, exposure, p0 = NULL, p1 = NULL, gamma = NULL,
  explanation = FALSE)
}
\arguments{
\item{mod}{Model object from \code{\link[=glm]{glm()}} or \code{\link[=coxph]{coxph()}}}

\item{exposure}{Character string. The name of the exposure variable.}

\item{p0}{Numeric. Estimated prevalence of the unmeasured confounder in the
unexposed population.}

\item{p1}{Numeric. Estimated prevalence of the unmeasured confounder in the
exposed population.}

\item{gamma}{Numeric. Estimated size of an unmeasured confounder.}

\item{explanation}{Logical. Indicates whether you'd like the function to output
a sentence explanation for inclusion in a manuscript.}
}
\value{
data frame
}
\description{
Tidy tipr
}
\examples{
library(magrittr)
## Fit a model
mod <- glm(vs ~ mpg, data = mtcars, family = binomial())

## Calculate p1
mod \%>\%
  tidy_tip(exposure = "mpg",
           p0 = 0,
           gamma = 1.4)

## Calculate p0
mod \%>\%
  tidy_tip(exposure = "mpg",
           p1 = 0,
           gamma = .71)

## Calculate gamma
mod \%>\%
  tidy_tip(exposure = "mpg",
           p0 = 0,
           p1 = 0.1)

## Calculate n
mod \%>\%
  tidy_tip(exposure = "mpg",
           p0 = 0,
           p1 = 0.1,
           gamma = 1.1)
}
