% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tip.R
\name{tip_b}
\alias{tip_b}
\title{Tip a result with a binary confounder.}
\usage{
tip_b(p1 = NULL, p0 = NULL, gamma = NULL, lb = NULL, ub = NULL,
  explanation = FALSE)
}
\arguments{
\item{p1}{estimated prevalence of the unmeasured confounder in the exposed population}

\item{p0}{estimated prevalence of the unmeasured confounder in the unexposed population}

\item{gamma}{estimated size of an unmeasured confounder}

\item{lb}{lower bound of your observed effect}

\item{ub}{upper bound of your observed effect}

\item{explanation}{logical value indicates whether you'd like the function to output a sentence explanation for inclusion in a manuscript.}
}
\description{
Choose two of the following three to specify, and the third will be estimated:
\itemize{
\item \code{p1}
\item \code{p0}
\item \code{gamma}
}
}
