% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tip.R
\name{tip_hr}
\alias{tip_hr}
\title{Tip an observed hazard ratio with a normally distributed confounder.}
\usage{
tip_hr(
  hr,
  smd = NULL,
  outcome_association = NULL,
  verbose = TRUE,
  hr_correction = FALSE
)
}
\arguments{
\item{hr}{Numeric positive value. Observed exposure - outcome hazard ratio.
This can be the point estimate, lower confidence bound, or upper
confidence bound.}

\item{smd}{Numeric. Estimated difference in scaled means between the
unmeasured confounder in the exposed population and unexposed population}

\item{outcome_association}{Numeric positive value. Estimated association
between the unmeasured confounder and the outcome}

\item{verbose}{Logical. Indicates whether to print informative message.
Default: \code{TRUE}}

\item{hr_correction}{Logical. Indicates whether to use a correction factor.
The methods used for this function are based on relative risks. For rare
outcomes, a hazard ratio approximates a relative risk. For common outcomes,
a correction factor is needed. If you have a common outcome (>15\%),
set this to \code{TRUE}. Default: \code{FALSE}.}
}
\value{
Data frame.
}
\description{
choose one of the following, and the other will be estimated:
\itemize{
\item \code{smd}
\item \code{outcome_association}
}
}
\examples{
## to estimate the association between an unmeasured confounder and outcome
## needed to tip analysis
tip_hr(1.2, smd = -2)

## to estimate the number of unmeasured confounders specified needed to tip
## the analysis
tip_hr(1.2, smd = -2, outcome_association = .99)

}
