% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tip_mod.R
\name{tip_mod}
\alias{tip_mod}
\title{Tidy tip a result with unmeasured confounding.}
\usage{
tip_mod(mod, exposure, rr_eu = NULL, rr_ud = NULL)
}
\arguments{
\item{mod}{Model object from \code{\link[=glm]{glm()}} or \code{\link[=coxph]{coxph()}}}

\item{exposure}{Character string. The name of the exposure variable.}

\item{rr_eu}{Numeric. The association between the exposure and the unmeasured confounder.}

\item{rr_ud}{Numeric. The association between the unmeasured confounder and the outcome.}
}
\value{
data frame
}
\description{
Tidy tip a result with unmeasured confounding.
}
\examples{
\dontrun{
## Fit a model
mod <- glm(vs ~ mpg, data = mtcars, family = binomial())

## Solve for the association between the exposure and unmeasured
## confounder needed to tip the analysis
mod \%>\%
 tip_mod(exposure = "mpg",
     rr_ud = 1.4)

## Solve for the association between the unmeasured confounder
## and the outcome needed to tip the analysis
mod \%>\%
 tip_mod(exposure = "mpg",
     rr_eu = 1.4)
## Solve for the minimum joint association between the
## exposure and unmeasured confounder and the unmeasured confounder
## and outcome, also known as the E-value
mod \%>\%
  tip_mod(exposure = "mpg")
}
}
