% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TIP_signature_symbol}
\alias{TIP_signature_symbol}
\title{cancer-immunity cycle related gene set}
\format{
An object of class \code{list} of length 28.
}
\source{
TIP web server \url{http://biocc.hrbmu.edu.cn/TIP/index.jsp}
}
\usage{
TIP_signature_symbol
}
\description{
A \code{list} of gene set vector get from TIP web server, all symbol
have checked use \code{HGNChelper} package.
}
\examples{
data("TIP_signature_symbol", package="TIPR")
head(TIP_signature_symbol)

}
\keyword{datasets}
