% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipr-package.R
\docType{package}
\name{tipr-package}
\alias{tipr}
\alias{tipr-package}
\title{tipr}
\description{
The tipr package.
}
\references{
D'Agostino McGowan, L, (2022). tipr: An R package for sensitivity analyses
for unmeasured confounders. Journal of Open Source Software, 7(77), 4495.

D’Agostino McGowan, L. (2022). Sensitivity Analyses for Unmeasured
Confounders. Current Epidemiology Reports, 9(4), 361-375.

VanderWeele, TJ, and Peng D (2017). Sensitivity Analysis in Observational
Research: Introducing the E-Value. Ann Intern Med, 167(4), 268–74.

Cinelli, C, & Hazlett, C (2020). Making sense of sensitivity: Extending
omitted variable bias. Journal of the Royal Statistical Society: Series B
(Statistical Methodology), 82(1), 39–67.

Lin, DY, Psaty, BM, & Kronmal, RA. (1998). Assessing the sensitivity
of regression results to unmeasured confounders in observational studies.
Biometrics, 54(3), 948–963.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/LucyMcGowan/tipr/issues}
}

}
\author{
\strong{Maintainer}: Lucy D'Agostino McGowan \email{lucydagostino@gmail.com} (\href{https://orcid.org/0000-0002-6983-2759}{ORCID})

Other contributors:
\itemize{
  \item Malcolm Barrett \email{malcolmbarrett@gmail.com} (\href{https://orcid.org/0000-0003-0299-5825}{ORCID}) [contributor]
}

}
\keyword{internal}
