% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tip_coef.R
\name{lm_tip}
\alias{lm_tip}
\title{Deprecated}
\usage{
lm_tip(effect, smd = NULL, outcome_effect = NULL, verbose = TRUE)
}
\arguments{
\item{effect}{Numeric. Observed exposure - outcome effect from a regression
model. This can be the beta coefficient, the lower confidence bound of
the beta coefficient, or the upper confidence bound of the beta
coefficient.}

\item{smd}{Numeric. Estimated scaled mean difference between the
unmeasured confounder in the exposed population and unexposed population}

\item{outcome_effect}{Numeric positive value. Estimated relationship
between the unmeasured confounder and the outcome}

\item{verbose}{Logical. Indicates whether to print informative message.
Default: \code{TRUE}}
}
\description{
This function has been deprecated and replaced by \code{\link{tip_coef}}.
}
