% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIP.R
\name{TIP}
\alias{TIP}
\title{Run TIP analysis}
\usage{
TIP(expression, perm = 100, seed = 1234, verbose = TRUE, ncores = 1)
}
\arguments{
\item{expression}{expression matrix with gene in row and sample in column.
You'd better use \code{log2(tpm+1)} normalized data, or preprocess your data with \code{preprocess}}

\item{perm}{permutation times, default 100}

\item{seed}{random state,default 1234}

\item{verbose}{Gives information about each calculation step. Default: TRUE.}

\item{ncores}{the parallel number of cores to run analysis, default 1.}
}
\value{
return a matrix with TIP activity score
}
\description{
Run TIP web server analysis
}
