% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipr-package.R
\docType{package}
\name{tipr}
\alias{tipr}
\alias{tipr-package}
\title{tipr}
\description{
tipr
}
\references{
Lin, D. Y., Psaty, B. M., & Kronmal, R. A. (1998). Assessing the sensitivity of regression results to unmeasured confounders in observational studies. Biometrics, 54(3), 948–963.
}
